requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-easy-pdf==0.1.1
django-utils-six==2.0
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
pdfkit==1.0.0
Pillow==9.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.5
